Now have a fully flegded lines class that works!
You can now instantiate a lines object and use it to play a game against the PC.
The previous bug (where it didn't think the empty sub-partition of [1,2] was a winning position) has been fixed.
There's also a little game_board_disply method that gives a pseudo graphical representation of the board!
Now onto the Tkinter front end ...