All dictionaries for classes now defined in init
Previously I was making the mistake of definig them in the class but not in the init method,
which had the effect that they were being shared through all class instances. This meant,
for example, that if you played two games (at the same time or one after the other - just
both fired from the same initial setup GUI), winning_moves from one might get mixed up with
winning_moves from the other, etc.  So this is a definite bug and a good fix!